fix system race condition warning [UAT-20]

+++ cli/index.js
@@ -39,33 +39,37 @@ if (!fs.existsSync("./fixtures/templates/docker-compose.md")) {
   return;
 }
 
-// Read the contents of the template file
-let template = fs.readFileSync(
-  "./fixtures/templates/docker-compose.md",
-  "utf8"
-);
+if (fs.existsSync("./docker-compose.yml")) {
+  console.log("docker-compose.yml already exists.");
+} else {
+  // Read the contents of the template file
+  let template = fs.readFileSync(
+    "./fixtures/templates/docker-compose.md",
+    "utf8"
+  );
 
-// Define default values
-let defaults = {
-  "#{SERVICE_NAME}": "udx-worker",
-  "#{USER}": "udx-worker",
-  "#{APP_PATH}": ".",
-};
+  // Define default values
+  let defaults = {
+    "#{SERVICE_NAME}": "udx-worker",
+    "#{USER}": "udx-worker",
+    "#{APP_PATH}": ".",
+  };
 
-// Override default values with command line arguments, if provided
-if (parsed.service_name) defaults["#{SERVICE_NAME}"] = parsed.service_name;
-if (parsed.user) defaults["#{USER}"] = parsed.user;
-if (parsed.app_path) defaults["#{APP_PATH}"] = parsed.app_path;
+  // Override default values with command line arguments, if provided
+  if (parsed.service_name) defaults["#{SERVICE_NAME}"] = parsed.service_name;
+  if (parsed.user) defaults["#{USER}"] = parsed.user;
+  if (parsed.app_path) defaults["#{APP_PATH}"] = parsed.app_path;
 
-// Replace variables in the template with their default values
-for (const variable in defaults) {
-  const value = defaults[variable];
-  const regex = new RegExp(variable, "g");
-  template = template.replace(regex, value);
+  // Replace variables in the template with their default values
+  for (const variable in defaults) {
+    const value = defaults[variable];
+    const regex = new RegExp(variable, "g");
+    template = template.replace(regex, value);
+  }
+
+  // Use the template to create a new docker-compose.yml file
+  fs.writeFileSync("./docker-compose.yml", template);
 }
-
-// Use the template to create a new docker-compose.yml file
-fs.writeFileSync("./docker-compose.yml", template);
 
 // Check if Docker containers are running
 exec("docker-compose ps", (error, stdout, stderr) => {
